Rename utask to task, rename signature to fnSignature
